More correct & efficient `AnyValue` decoding
This reworks the decoding machinery in `AnyValue` to be more efficient and produce the correct results consistently.  All other changes were required to support this change.

* Attempts to decode as keyed, then unkeyed and then single valued; utilizing the decoding machinery better
* Added “raw value” decoding enhancement when trying as a single value. This only works when it is utilizing a PC based decoder; in all other cases it falls back to trying each value in turn
* Adds unwrapping as a requirement to `Value`. Now any intermediate value can be unwrapped to native Swift/Foundation primitives
* JSON Numbers now use a specialized type that allows reproducible reading and writing

## JSON Numbers
Previously JSON numbers were stored as `Float80` it was decided that this doesn’t always honor what was encoded in JSON. When fixing the `AnyValue` decoding a method was needed to produce a value consistent with the encoded JSON and now a `JSON.Number` type is used that stores the parsed text internally ensuring we can always serialize what was parsed and that conversion to the target type is always from the parsed text.