Use `ZonedDate` for ASN.1 GeneralizedTime
This replaces the use of `Foundation.Date` with a custom `ZonedDate` type. This allows supporting properly reading/writing GeneralizedTime values with a specific time zone.

This also uses specialized parsers for reading GeneralizedTime values that properly allows reading values with optional fractional seconds and optional zones; fixing #13.

This will require a major version bump as it changes the public API.